refactor(agents): split CLI harnesses into subpackages
Move each CLI agent into its own subpackage that separates the run driver
(agent) from the trajectory parser (parsing), and rename the Gemini harness
module to gemini_cli to distinguish it from the Gemini model client.
