+++ devops_bench/agents/cli/gemini_cli/agent.py
@@ -15,10 +15,8 @@
 """Gemini CLI agent harness driving the ``gemini`` binary.
 
 Trajectory extraction reads the official ``--output-format stream-json`` event
-stream from stdout — no ``~/.gemini/tmp/...`` disk reads, no session-id glob,
-no internal-schema parsing. ``tool_use``/``tool_result`` events fold into the
-canonical :class:`ToolCall` list; ``result`` events carry the final text and
-the aggregated token usage.
+stream from stdout (see :mod:`~devops_bench.agents.cli.gemini_cli.parsing`) — no
+``~/.gemini/tmp/...`` disk reads, no session-id glob, no internal-schema parsing.
 
 Capability wiring: the allowed-tools list comes from the aggregated
 ``config.capabilities.allowed_tools`` (across every bound MCP server) and the
@@ -29,24 +27,24 @@ native context-file mechanism).
 
 from __future__ import annotations
 
-import json
 import os
 import tempfile
 from pathlib import Path
 
 from devops_bench.agents.base import AGENTS, AgentHarness
+from devops_bench.agents.cli.gemini_cli.parsing import parse_stream_json
 from devops_bench.agents.config import AgentConfig
-from devops_bench.agents.result import AgentResult, ToolCall
+from devops_bench.agents.result import AgentResult
 from devops_bench.core import SubprocessError, get_logger
 from devops_bench.core.subprocess import run
 
-__all__ = ["GeminiCliAgent", "parse_stream_json"]
+__all__ = ["GeminiCliAgent"]
 
 # Filename the Gemini CLI auto-loads from its working directory as the
 # operator brief / startup context (its native equivalent of a system prompt).
 _GEMINI_RULES_FILE = "GEMINI.md"
 
-_log = get_logger("agents.cli.gemini")
+_log = get_logger("agents.cli.gemini_cli")
 
 
 def _build_argv(target: str, prompt: str, allowed_tools: tuple[str, ...]) -> list[str]:
@@ -108,116 +106,6 @@ def _build_env(config: AgentConfig) -> dict[str, str]:
     if config.extra_env:
         overlay.update(config.extra_env)
     return overlay
-
-
-def parse_stream_json(stdout: str) -> tuple[str, list[dict], dict, list[str]]:
-    """Parse a Gemini ``--output-format stream-json`` stdout stream.
-
-    The stream is newline-delimited JSON events. The parser is intentionally
-    lenient (an unknown event type is skipped) and surfaces both per-line JSON
-    decode errors and unmatched ``tool_result`` events on the ``errors`` list
-    rather than silently dropping them.
-
-    | Event type      | Fields read                                              |
-    |-----------------|---------------------------------------------------------|
-    | ``init``        | (ignored)                                               |
-    | ``message``     | ``role`` (assistant text accumulated into the output)   |
-    | ``tool_use``    | ``tool_name``, ``tool_id``, ``parameters``              |
-    | ``tool_result`` | ``tool_id``, ``status`` (no payload in the stream)      |
-    | ``error``       | recorded on the errors list                             |
-    | ``result``      | ``stats`` (token usage); terminal status                |
-
-    Args:
-        stdout: Raw process stdout, possibly empty.
-
-    Returns:
-        A ``(output, trajectory, tokens, errors)`` tuple. ``trajectory`` is a
-        list of ``ToolCall.to_dict()`` mappings ordered as emitted.
-    """
-    output_parts: list[str] = []
-    tokens: dict = {}
-    errors: list[str] = []
-    pending: dict[str, ToolCall] = {}
-    trajectory: list[ToolCall] = []
-
-    for lineno, raw in enumerate(stdout.splitlines(), start=1):
-        line = raw.strip()
-        if not line:
-            continue
-        try:
-            event = json.loads(line)
-        except json.JSONDecodeError as exc:
-            errors.append(f"stream-json line {lineno} parse error: {exc}")
-            continue
-        if not isinstance(event, dict):
-            continue
-
-        etype = event.get("type")
-        if etype == "message":
-            # ``role="user"`` echoes the prompt and is skipped.
-            if event.get("role") in ("assistant", "model"):
-                content = event.get("content")
-                if isinstance(content, str):
-                    output_parts.append(content)
-                elif isinstance(content, list):
-                    # Defensive: a parts-list shape (``[{"text": ...}]``).
-                    for part in content:
-                        if isinstance(part, dict) and isinstance(part.get("text"), str):
-                            output_parts.append(part["text"])
-        elif etype == "tool_use":
-            call_id = event.get("tool_id") or event.get("id") or event.get("tool_use_id") or ""
-            args = event.get("parameters")
-            if args is None:
-                args = event.get("input")
-            if args is None:
-                args = event.get("args")
-            call = ToolCall(
-                name=event.get("tool_name") or event.get("name", ""),
-                args=args if isinstance(args, dict) else {},
-                status="called",
-            )
-            trajectory.append(call)
-            if call_id:
-                pending[str(call_id)] = call
-        elif etype == "tool_result":
-            call_id = event.get("tool_id") or event.get("tool_use_id") or event.get("id") or ""
-            target = pending.pop(str(call_id), None) if call_id else None
-            if target is None:
-                errors.append(f"stream-json tool_result without matching tool_use (id={call_id!r})")
-                continue
-            # tool_result carries only a status; accept a content/output
-            # payload as a fallback when present.
-            content = event.get("content")
-            if content is None:
-                content = event.get("output")
-            if content is not None:
-                target.result = content if isinstance(content, str) else json.dumps(content, default=str)
-            status = str(event.get("status", "")).lower()
-            failed = bool(event.get("is_error")) or status in ("error", "failed", "failure")
-            target.status = "error" if failed else "completed"
-        elif etype == "error":
-            msg = event.get("message") or event.get("error") or str(event)
-            errors.append(f"stream-json error event: {msg}")
-        elif etype == "result":
-            # Terminal event: answer streams via ``message`` events and token
-            # usage rides under ``stats``; accept ``output``/``response`` and
-            # ``tokens``/``usage`` as fallbacks.
-            tail = event.get("output") or event.get("response")
-            if isinstance(tail, str) and tail:
-                output_parts.append(tail)
-            stats = event.get("stats")
-            usage = event.get("tokens") or event.get("usage")
-            if isinstance(stats, dict):
-                tokens = {
-                    "input": stats.get("input_tokens"),
-                    "output": stats.get("output_tokens"),
-                    "total": stats.get("total_tokens"),
-                    "cached": stats.get("cached"),
-                }
-            elif isinstance(usage, dict):
-                tokens = usage
-
-    return "".join(output_parts), [call.to_dict() for call in trajectory], tokens, errors
 
 
 @AGENTS.register("gemini")
