+++ devops_bench/agents/cli/openclaw/__init__.py
@@ -12,11 +12,16 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-"""Concrete CLI-binary agent harnesses (``gemini``, ``oc``).
+"""OpenClaw CLI agent harness driving the local ``oc`` binary.
 
-Each harness lives in its own subpackage (:mod:`.gemini_cli`, :mod:`.openclaw`)
-that splits the run driver (``agent``) from the trajectory parsers (``parsing``)
-and self-registers under its canonical key on import via ``@AGENTS.register``.
-Subpackages are imported by the harness at call time, not by this package's
-``__init__`` — keeping the import graph light.
+The harness driver lives in :mod:`.agent` and the trajectory-bundle parsers in
+:mod:`.parsing`. Importing this package self-registers the agent under the
+``"openclaw"`` key via ``@AGENTS.register``.
 """
+
+from __future__ import annotations
+
+from devops_bench.agents.cli.openclaw.agent import OpenClawAgent
+from devops_bench.agents.cli.openclaw.parsing import parse_trajectory_export
+
+__all__ = ["OpenClawAgent", "parse_trajectory_export"]
